+++ CHANGELOG.md
@@ -25,6 +25,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Added `from_polyline` to `compas_timber.elements.Wall.
 * Added `WallJoint` to `compas_timber.connections`.
 * Added error handling when BTLx processing from geometry fails in GH.
+* Added new `Slab` class to `compas_timber.elements`.
+* Added `Slab` GH component.
 
 ### Changed